feat: remove controllers
